+++ Source/Agents/Players/Mpg123/Mpg123Worker.cs
@@ -146,11 +146,11 @@ namespace Polycode.NostalgicPlayer.Agent.Player.Mpg123
 				return false;
 			}
 
-			mpg123Handle.Mpg123_Rates(out int[] supportedRates, out ulong number);
+			mpg123Handle.Mpg123_Rates(out long[] supportedRates, out ulong number);
 			if (number > 0)
 			{
 				// Set the output format to 32-bit on every rate
-				foreach (int rate in supportedRates)
+				foreach (long rate in supportedRates)
 				{
 					result = mpg123Handle.Mpg123_Format(rate, Mpg123_ChannelCount.Mono | Mpg123_ChannelCount.Stereo, Mpg123_Enc_Enum.Enc_Signed_32);
 					if (result != Mpg123_Errors.Ok)
@@ -329,7 +329,7 @@ namespace Polycode.NostalgicPlayer.Agent.Player.Mpg123
 		/// Return the frequency the sample is stored with
 		/// </summary>
 		/********************************************************************/
-		public override int Frequency => frameInfo.Rate;
+		public override int Frequency => (int)frameInfo.Rate;
 
 
 
@@ -456,6 +456,7 @@ namespace Polycode.NostalgicPlayer.Agent.Player.Mpg123
 							break;
 						}
 					}
+
 					break;
 				}
 
@@ -517,6 +518,7 @@ namespace Polycode.NostalgicPlayer.Agent.Player.Mpg123
 							break;
 						}
 					}
+
 					break;
 				}
