+++ Source/NostalgicPlayer.Kit.C/CTypes.cs
@@ -6,13 +6,13 @@
 global using c_char = System.SByte;
 global using c_short = System.Int16;
 global using c_int = System.Int32;
-global using c_long = System.Int32;
+global using c_long = System.Int64;
 global using c_long_long = System.Int64;
 
 global using c_uchar = System.Byte;
 global using c_ushort = System.UInt16;
 global using c_uint = System.UInt32;
-global using c_ulong = System.UInt32;
+global using c_ulong = System.UInt64;
 global using c_ulong_long = System.UInt64;
 
 global using c_float = System.Single;
